update 1 AM

+++ src/admin/TrainingSection.js
@@ -5,6 +5,8 @@ import axios from 'axios';
 const TrainingSection = () => {
     const [programs, setPrograms] = useState([]);
     const [sessions, setSessions] = useState([]);
+    const [groupsession, setGroupSession] = useState([]);
+    const [allSession, setAllSession] = useState([]);
     const navigate = useNavigate();
   
     useEffect(() => {
@@ -13,19 +15,36 @@ const TrainingSection = () => {
           const response = await axios.get('http://localhost:8080/api/session', {
             headers: { 'Content-Type': 'application/json' },
         });
-        console.log(response.data)
         setSessions(response.data)
         } catch (error) {
           
         }
       }
+      const fetchGroupSession = async () =>{
+        try {
+          const response = await axios.get('http://localhost:8080/api/groupsessions', {
+            headers: { 'Content-Type': 'application/json' },
+        });
+        setGroupSession(response.data)
+        } catch (error) {
+          
+        }
+      }
       fetchSession()
+      fetchGroupSession()
+      // setAllSession(...sessions,...groupsession)
       document.body.classList.add('programM-page');
   
       return () => {
         document.body.classList.remove('programM-page');
       };
     }, []);
+
+    useEffect(() => {
+      if (sessions.length > 0 && groupsession.length > 0) {
+          setAllSession([...sessions, ...groupsession]);
+      }
+  }, [sessions, groupsession]);
   
     const handleAddProgram = () => {
       navigate('/admin-trainingS-add');
@@ -39,21 +58,39 @@ const TrainingSection = () => {
       navigate('/admin-training-edit', { state: { programIndex: index } });
     };
   
-    const handleDeleteProgram = async (id) => {
+    const handleDeleteProgram = async (id,type) => {
       console.log(id)
       const confirmDelete = window.confirm("Are you sure you want to delete this program?");
     if (confirmDelete) {
-      try {
-        await axios.delete(`http://localhost:8080/api/session/${id}`, {
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        });
-
-        // setSessions(sessions.filter((program) => program.id !== id)); // Update programs list
-        alert('ลบข้อมูล Session เรียบร้อย');
-      } catch (error) {
-        console.error("Error deleting Session:", error);
+      if(type == "ส่วนตัว"){
+        try {
+          await axios.delete(`http://localhost:8080/api/session/${id}`, {
+            headers: {
+              'Content-Type': 'application/json',
+            },
+          });
+  
+          // setSessions(sessions.filter((program) => program.id !== id)); // Update programs list
+          alert('ลบข้อมูล Session เรียบร้อย');
+        } catch (error) {
+          console.error("Error deleting Session:", error);
+          alert('ไม่สามารถลบ Session ได้');
+        }
+      }if (type == "กลุ่ม") {
+        try {
+          await axios.delete(`http://localhost:8080/api/groupsessions/${id}`, {
+            headers: {
+              'Content-Type': 'application/json',
+            },
+          });
+  
+          // setSessions(sessions.filter((program) => program.id !== id)); // Update programs list
+          alert('ลบข้อมูล Session เรียบร้อย');
+        } catch (error) {
+          console.error("Error deleting Session:", error);
+          alert('ไม่สามารถลบ Session ได้');
+        }
+      } else {
         alert('ไม่สามารถลบ Session ได้');
       }
     }
@@ -88,7 +125,10 @@ return(
       </button>
       </div>
       <div className="program-list">
-        {sessions.length === 0 ? (
+        {console.log(sessions)}
+        {console.log(groupsession)}
+        {console.log(allSession)}
+        {allSession.length === 0 ? (
           <p>No training programs available.</p>
         ) : (
           <table>
@@ -96,7 +136,7 @@ return(
               <tr>
                 {/* <th>Image</th> */}
                 {/* <th>Program Name</th> */}
-                <th>Session ID</th>
+                {/* <th>Session ID</th> */}
                 <th>ชื่อโปรแกรม</th>
                 <th>ประเภท</th>
                 <th>ตารางเวลา</th>
@@ -105,10 +145,8 @@ return(
               </tr>
             </thead>
             <tbody>
-              {sessions.map((session, index) => (
+              {allSession.map((session, index) => (
                 <tr key={index}>
-                  {/* Session ID */}
-                  <td>{session.sessionId}</td> {/* Displaying session ID */}
 
                   {/* Program Name */}
                   <td>{session.program.programName}</td>
@@ -119,7 +157,7 @@ return(
                   {/* Schedule */}
                   <td>
                     {new Date(session.dateSession.sdate).toLocaleDateString('th-TH')}<br />
-                    {session.dateSession.startTime} - {session.dateSession.endTime}
+                    {session.dateSession.startTime.substring(0, 5)} - {session.dateSession.endTime.substring(0, 5)}
                   </td>
 
                   {/* Description */}
@@ -127,7 +165,7 @@ return(
 
                   {/* Actions */}
                   <td>
-                    <button className="delete" onClick={() => handleDeleteProgram(session.sessionId)}>Delete</button>
+                    <button className="delete" onClick={() => handleDeleteProgram(session.sessionId,session.program.programType)}>Delete</button>
                   </td>
                 </tr>
               ))}
